Moved helper code to graphQL folder
