+++ src/App.ts
@@ -13,53 +13,53 @@ import { importSchema } from 'graphql-import';
 import resolvers from './graphql/resolvers'
 import { ReqContext } from './graphql/types/server.types';
 import { peopleFromHouseholdLoader, householdLoader } from './graphql/loader'
-import { validateToken } from './helpers/account';
+import { validateToken } from './graphql/helpers/account';
 
 export const init = async () => {
-    /*
-    AWS.config.update({ region: 'us-east-2' });
-    const logger = winston.createLogger({
-        transports: [new WinstonCloudWatch({ logGroupName: 'AccessManagementStage', logStreamName: 'API' })],
-        format: winston.format.json()
-    });
-    logger.info("App Logger initialized");*/
+  /*
+  AWS.config.update({ region: 'us-east-2' });
+  const logger = winston.createLogger({
+      transports: [new WinstonCloudWatch({ logGroupName: 'AccessManagementStage', logStreamName: 'API' })],
+      format: winston.format.json()
+  });
+  logger.info("App Logger initialized");*/
 
-    dotenv.config();
-    const container = new Container();
-    await container.loadAsync(bindings);
-    const app = new InversifyExpressServer(container, null, null, null, CustomAuthProvider);
+  dotenv.config();
+  const container = new Container();
+  await container.loadAsync(bindings);
+  const app = new InversifyExpressServer(container, null, null, null, CustomAuthProvider);
 
 
-    const configFunction = (expApp: express.Application) => {
-        // expApp.use(bodyParser({ limit: "50mb" }));
-        // expApp.use()
-        expApp.use(bodyParser.urlencoded({ extended: true }));
-        expApp.use(bodyParser.json({ limit: "50mb" }));
-        expApp.use(cors())
-    };
+  const configFunction = (expApp: express.Application) => {
+    // expApp.use(bodyParser({ limit: "50mb" }));
+    // expApp.use()
+    expApp.use(bodyParser.urlencoded({ extended: true }));
+    expApp.use(bodyParser.json({ limit: "50mb" }));
+    expApp.use(cors())
+  };
 
-    const server = app.setConfig(configFunction).build();
-    const graphQLServer = new ApolloServer({
-      typeDefs: importSchema('src/graphql/schema/schema.graphql'),
-      resolvers,
-      validationRules: [depthLimit(5)],
-      context: (ctx: ReqContext) => {
-        const { authorization } = ctx.req.headers
-        if (authorization && typeof authorization === 'string' && authorization.startsWith('Bearer ')) {
-          const token = authorization.substr(7)
-          const me = validateToken(token)
-          ctx.me = me
-        }
-        ctx.peopleFromHouseHoldLoader = peopleFromHouseholdLoader()
-        ctx.householdLoader = householdLoader()
+  const server = app.setConfig(configFunction).build();
+  const graphQLServer = new ApolloServer({
+    typeDefs: importSchema('src/graphql/schema/schema.graphql'),
+    resolvers,
+    validationRules: [depthLimit(5)],
+    context: (ctx: ReqContext) => {
+      const { authorization } = ctx.req.headers
+      if (authorization && typeof authorization === 'string' && authorization.startsWith('Bearer ')) {
+        const token = authorization.substr(7)
+        const me = validateToken(token)
+        ctx.me = me
+      }
+      ctx.peopleFromHouseHoldLoader = peopleFromHouseholdLoader()
+      ctx.householdLoader = householdLoader()
 
-        return ctx
-      },
-    })
+      return ctx
+    },
+  })
 
-    graphQLServer.applyMiddleware({
-      app: server,
-      path: '/graphql',
-    })
-    return server;
+  graphQLServer.applyMiddleware({
+    app: server,
+    path: '/graphql',
+  })
+  return server;
 }
