+++ src/graphql/resolvers/group.resolver.ts
@@ -2,7 +2,7 @@ import _ from 'lodash'
 import { UserInputError } from 'apollo-server'
 import { prisma } from '../prisma'
 import { QueryGroupsArgs, ReqContext, SortDirection, QueryGroupArgs, GroupsResult, Group } from '../types'
-import { initPagination } from '../../helpers'
+import { initPagination } from '../helpers'
 
 export default {
   Query: {
@@ -11,9 +11,11 @@ export default {
       if (!id) {
         throw new UserInputError('group_id is required')
       }
-      const group = await prisma.groups.findFirst({ where: {
-        id
-      }})
+      const group = await prisma.groups.findFirst({
+        where: {
+          id
+        }
+      })
       return group
     },
     groups: async (root: any, args: QueryGroupsArgs, ctx: ReqContext): Promise<Group[] | null> => {
